Abstracting Post to global types
